refactor number of secondaries logic
